Fix YAML syntax of workflow file
YAML should be wiped off the face of the Earth. Jesus Christ, if you're going to force your syntax on everyone, at least force a good syntax, not whatever the hell this is. Anyone who uses 2-line indentation needs to slap themselves on the wrist until they learn to indent properly, with 4 spaces.